requirements.txt
streamlit
pandas
plotly
requests
beautifulsoup4
psycopg2-binary
